+++ test/tool/device_display_model/fixtures/fake_security.sh
@@ -1,6 +1,6 @@
-#!/bin/zsh
+#!/usr/bin/env bash
 
-print -r -- "<?xml version=\"1.0\" encoding=\"UTF-8\"?>
+printf '%s\n' "<?xml version=\"1.0\" encoding=\"UTF-8\"?>
 <plist version=\"1.0\"><dict>
 <key>Entitlements</key><dict>
 <key>application-identifier</key><string>TEAM.${OMF_FAKE_IOS_PROFILE_BUNDLE:-dev.ventairy.oh-my-flutter.device-display-collector}</string>